+++ CHANGELOG.md
@@ -106,17 +106,16 @@ In `FlexColorScheme` and `FlexThemeData` light/dark constructors, the `usedColor
 - Added theming of indicator border radius to `NavigationBar` to panel **Navigation Bar**.
 - Added theming of indicator border radius to `NavigationRail` to panel **Navigation Rail**.
 - Added theming of action color to `SnackBar` action button to panel **BottomSheet, Snack Banner**. The UI also got better testing feature of both fixed and floating `SnackBar` and correct defaults info for different settings and modes.
-- Added controls for theming menus consistently to the **Menu** panel.
+- Added controls to theme menus consistently to the **Menu** panel.
 
 **TASKS DONE**
 - Added value tests for all new color definitions.
 
 **TODO BEFORE FCS BETA 7.0.0-dev.3 RELEASE**
 
+- WIP ADD: Page with pre-made example theme configurations. Panel added, configs WIP.
 - TODO: TintedDisabled and TintedEffect, needs review and implementations.
 - TODO: Check status of issue https://github.com/flutter/flutter/issues/100027 in Flutter 3.7.1
-- MAYBE ADD: Page with pre-made example theme configurations.
-- MAYBE ADD: TextField settings examples.
 
 **TODO BEFORE FCS STABLE 7.0 RELEASE**